requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
simple-websocket==1.1.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
requests
psycopg2
python-dotenv